Major update for external input controls #348
- Rewrite of the input event mechanism
- Consistent focus/frame/glow logic
- XKeyPad enhancements for navigation, glow

+++ src/elem/XRadial.h
@@ -225,7 +225,7 @@ bool gslc_ElemXRadialDrawRadial(gslc_tsGui* pGui,gslc_tsElemRef* pElemRef,gslc_t
 
 #define gslc_ElemXRadialCreate_P(pGui,nElemId,nPage,nX,nY,nW,nH,\
     nMin_,nMax_,nVal_,colFrame_,colFill_,colGauge_) \
-  static const uint8_t nFeatures##nElemId = GSLC_ELEM_FEA_VALID | \
+  static const uint16_t nFeatures##nElemId = GSLC_ELEM_FEA_VALID | \
     GSLC_ELEM_FEA_GLOW_EN | GSLC_ELEM_FEA_FILL_EN; \
   static gslc_tsXRadial sGauge##nElemId;                           \
   sGauge##nElemId.nMin = nMin_;                                   \
@@ -274,7 +274,7 @@ bool gslc_ElemXRadialDrawRadial(gslc_tsGui* pGui,gslc_tsElemRef* pElemRef,gslc_t
 
 #define gslc_ElemXRadialCreate_P(pGui,nElemId,nPage,nX,nY,nW,nH,\
     nMin_,nMax_,nVal_,colFrame_,colFill_,colGauge_) \
-  static const uint8_t nFeatures##nElemId = GSLC_ELEM_FEA_VALID | \
+  static const uint16_t nFeatures##nElemId = GSLC_ELEM_FEA_VALID | \
     GSLC_ELEM_FEA_GLOW_EN | GSLC_ELEM_FEA_FILL_EN; \
   static gslc_tsXRadial sGauge##nElemId;                           \
   sGauge##nElemId.nMin = nMin_;                                   \
